convert "fisher" method of finding polynomial
Since opencv doesn't have a way to solve generalized eigenvalue problem - had use Eigen for that. Currently settled with QZ decomposition, since it seems to be used in Matlab for badly conditioned matrices. Eigen doesn't calculate eigenvectors during QZ decomposition, so simplest approach was taken - solve for the eigenvector using SVD, given A, B and corresponding eigenvalue. Maybe something more involved should be used.

The data used in find_polynomials tests seemed to be badly conditioned and the results differed between Matlab and C++, so had to use another input for testing. Or maybe this is because of some bug. Should do more testing.